feat(runloop) rebuild plugins_iterator and router via a timer
This change creates a timer which checks and rebuilds the plugins iterator and router, once per second, so the rebuilding is partially uncoupled from receiving a request.

Extracted from #4488, with some minor differences:

* `rebuild_sync` was renamed to `safe_rebuild`
* Some functions have changed so that they return nil, err instead of
logging the error and just returning false
* The `rebuild` utility function now returns nil, err as well, it is up
to the caller to decide how to deal with errors. However the function
still logs a couple things on its own, when fallbacks are used.

Co-Authored-By: Aapo Talvensaari <bungle@konghq.com>