refactor(runloop) make router functions similar to plugins_iterator's
This is to align functionality with plugins_iterator.

* `get_xxx` is just a getter. It has no functionality and cannot fail
* `build_xxx` is a setter. It always attempts to build a new
  iterator/router and replace the current one
* `update_xxx` is a "safe setter". It calls `build_xxx` only after
  checking that the versions are appropiate and after checking for
  concurrency

The previous get_router did more than just get (it was a mixture of
update + get).

The previous check_router_rebuild is an almost direct rename to
update_router.

build_router only needed a small change: it only sets the router if
there are no errors (notice the use of `new_router`).

The tests have been rewritten so that they spy on kong.cache.get
instead of having to export a private function in order to test
semaphores.