perf(runloop) detect & retry version changes while rebuilding
Both the Router and the Plugins Iterator are built, at their core, by
looping over database registers: routes in the former and plugins in the
later.

This iteration happens by getting them from the database in batches of
1000. While the db request happens, the worker is waiting for io and
other tasks may enter, including registering events of router / plugin
updates. This change detects this change and fails the current build.

However, the change also includes a retry mechanism (implemented
differently in #4488, with recursion instead of a loop). If a rebuild
fails, it will be retried up to 6 times, with sleeps in-between.